Simplify runJob
Move more repetitive code to JobUtils
